Created the room and positioning
Created a room with a TV and couch.
The TV has a canvas that displays text.

The beginning text asks the user to stand on a blue square and then the calibration proceeds to the left arm (not completed).

The confirmation button is either trigger or the space (just in case the patient can move their fingers)