Started data system
The values that hold the data for the maximum/minimum arm threshold taken from the cube.

The way I am tackling this problem is with two objects a leftParent object with the child being the left cube the player moves.

A float value holds the distance on 1 axis (x for left, z for forward, y for top and bottom) between the two object and then divides it by the expected maximum value (for example the fathest distance the left block can get in a normal human scenanio is roughly 0.55f) so if the player has moved the cube a distance of 0.3 away from the parent then the total value will be (0.3/0.55) = 0.545 * 100 or 55% with makes sense since 0.55/0.55 is 100% and 0/0.55 is 0%. This pretty much normalized the data.

I also cleaned up the code a little and removed unnessacary comments