Finished normalizing data
Completed the normalizing process
Now (for both hands) the data is normalized from -1 to 1 for the back, front, left, right, top, and bottom cubes. I converted the values into a percent (0 to 100%) but it can be changed by removing the (* 100)  on each value. I thought it would look cleaner than a bunch of decimals. The process is still a little buggy. Your arm length plays a large part in how far you can push the cubes so updating the max values based on the players arm length could be something to look into