Saving, clamping, and cleaning
The values for each hand/axis are now saved in the player prefs.
Since playerPrefs are stored in the registry theoretically they can be accessed from another project but If we are combining the calibration system to the games it would be in the same project and we wouldn't need to worry about that. So for right now the save is only one way (the info is not used) I also clamped the cube data to the min and max values (-1, 0; and 0, 1) so you can push a cube to the wrong side. I also removed Debug.Log comments and other unnessacary comments.