Udjusted height
Created two variables: heightInFeet and sitting (might remove). Depending on the patients height in feet (eg: 6ft) calculates it in meters for unity units and sets the camera height.
If sitting is enabled divides that value by 1.3. However, this might be obsolete since you can just sit down and it will lower the camera automatically.